refactor: Clean up Home component
 - GenerateSchedules service
 - matrixTemplate removed to Schedule context
 - Add ServerStatus component
