+++ src/components/Banner/Banner.jsx
@@ -4,9 +4,9 @@ import { BannerWrapper } from './Banner.styles'
 function Banner() {
   return (
     <BannerWrapper>
-      Coded with ♡ by <a href="https://github.com/sjdonado" target="_blank">Juan Rodriguez,</a>
-      <a href="https://github.com/rmacuna" target="_blank">Roberto Acuña,</a>
-      <a href="https://github.com/krthr" target="_blank"> y Wilson Tovar</a>
+      Coded with ♡ by <a href="https://github.com/sjdonado" target="_blank" rel="noopener noreferrer">Juan Rodriguez,</a>
+      <a href="https://github.com/rmacuna" target="_blank" rel="noopener noreferrer">Roberto Acuña,</a>
+      <a href="https://github.com/krthr" target="_blank" rel="noopener noreferrer"> y Wilson Tovar</a>
     </BannerWrapper>
   )
 }
