+++ src/context/schedules-context.js
@@ -1,11 +1,9 @@
 import React from 'react';
 
 const SchedulesContext = React.createContext({
-  matrixTemplate: [],
   schedules: [],
   currentSchedule: {},
   setCurrentSchedule: () => { },
-  setMatrixTemplate: () => { }
 });
 
 export default SchedulesContext;
