+++ src/pages/home/Home.jsx
@@ -1,8 +1,6 @@
-import React, { useState } from 'react'
+import React, { useState, useMemo, useEffect } from 'react'
 import { Row, Col } from 'react-flexbox-grid'
-import { useQuery } from '@apollo/react-hooks'
 
-import { GET_SERVER_STATUS_QUERY } from '../../graphql/queries'
 
 import {
   ModalSubtitle,
@@ -24,6 +22,7 @@ import {
 import Modal from '../../components/Modal/Modal'
 import Sidenav from '../../components/Sidenav/Sidenav'
 import SearchSelect from './SearchSelect/SearchSelect'
+import ServerStatus from './ServerStatus/ServerStatus'
 import Paginator from './Paginator/Paginator'
 
 import SubjectDetails from './SubjectDetails/SubjectDetails'
@@ -36,7 +35,11 @@ import SchedulesContext from '../../context/schedules-context'
 import Banner from '../../components/Banner/Banner'
 import BlockCheckbox from '../../components/Checkbox/Checkbox'
 
+import generateSchedules from '../../services/generateSchedules'
+
 import $ from 'jquery'
+
+const EMPTY_MATRIX = generateEmptyMatrix()
 
 const DEFAULT_EMPTY_SCHEDULE = {
   matrix: null,
@@ -49,19 +52,26 @@ function Home() {
   const [localSubjects, setLocalSubjects] = useState([])
   const [localSchedules, setLocalSchedules] = useState([])
   const [allowFullGroups, setAllowFullGroups] = useState(false)
-  const [localMatrixTemplate, setLocalMatrixTemplate] = useState(generateEmptyMatrix())
+  const [localMatrixTemplate, setLocalMatrixTemplate] = useState(EMPTY_MATRIX)
   const [localCurrentSchedule, setLocalCurrentSchedule] = useState(DEFAULT_EMPTY_SCHEDULE)
 
-  const ServerStatus = () => {
-    const { loading, error, data } = useQuery(GET_SERVER_STATUS_QUERY)
-    if (loading || error) return null
+  // const schedules = useMemo(() => generateSchedules(localSubjects, localMatrixTemplate, allowFullGroups),
+  //   [generateSchedules, localSubjects, localMatrixTemplate, allowFullGroups])
 
-    const { totalGroups, updatedAt } = data.getServerStatus;
-    return (
-      <span className="server_status">
-        Última actualización: {new Date(updatedAt).toLocaleString()} - Grupos obtenidos: {totalGroups}
-      </span>
-    )
+  // useEffect(() => {
+  //   loadSchedules()
+  // })
+
+  const loadSchedules = (subjects = localSubjects, matrixTemplate = localMatrixTemplate) => {
+    const schedules = generateSchedules(subjects, matrixTemplate, allowFullGroups)
+    if (schedules.length) {
+      setLocalCurrentSchedule(schedules[0])
+      setLocalSchedules(schedules)
+    } else {
+      setLocalCurrentSchedule(DEFAULT_EMPTY_SCHEDULE)
+      setCurrentPage(0)
+    }
+    setIsLoading(false)
   }
 
   const [modal, setModal] = useState({
@@ -84,129 +94,20 @@ function Home() {
     })
   }
 
-  const parseDay = day => {
-    switch (day) {
-      case 'M':
-        return 0
-      case 'T':
-        return 1
-      case 'W':
-        return 2
-      case 'R':
-        return 3
-      case 'F':
-        return 4
-      case 'S':
-        return 5
-      default:
-        return 6
-    }
-  }
-
-  const addToScheduleMatrix = (group, matrix, matrixTemplate) => {
-    let days, start, end
-    let newMatrix = [...matrix.map(elem => [...elem])]
-    for (let schedule of group.schedule) {
-      days = schedule.day.split('')
-      while (days.length) {
-        start = Number(schedule.time.start.substring(0, 2))
-        end = Number(schedule.time.end.substring(0, 2))
-        for (let hour = start; hour < end; hour++) {
-          if (newMatrix[parseDay(days[0])][hour - 6]
-            || (matrixTemplate && matrixTemplate[parseDay(days[0])][hour - 6])) {
-            return false
-          }
-          newMatrix[parseDay(days[0])][hour - 6] = group.subject.name
-        }
-        days.shift()
-      }
-    }
-    return newMatrix
-  }
-
-  const generateSchedules = (
-    newLocalSubjects = localSubjects,
-    matrixTemplate = localMatrixTemplate,
-  ) => {
-    setLocalCurrentSchedule(DEFAULT_EMPTY_SCHEDULE)
-    setCurrentPage(0)
-
-    let groups = []
-    for (let subject of newLocalSubjects) {
-      groups = groups.concat(subject.groups)
-    }
-
-    const schedules = []
-    const groupKeys = new Set()
-    let newMatrix, groupKey, groupsToCompare
-
-    for (let group of groups) {
-      if (group.blocked || (!allowFullGroups && group.quota.free === 0)) {
-        continue
-      }
-      groupsToCompare = [...groups]
-      while (groupsToCompare.length) {
-        let subjectsIdsUsed = new Set()
-        let schedule = {
-          matrix: generateEmptyMatrix(),
-          groups: [],
-        }
-
-        newMatrix = addToScheduleMatrix(group, schedule.matrix, matrixTemplate)
-        if (!newMatrix) {
-          break
-        }
-        schedule.matrix = newMatrix
-        schedule.groups.push(group)
-        subjectsIdsUsed.add(group.subject.id)
-
-        if (groupsToCompare[0].nrc === group.nrc) {
-          groupsToCompare.shift()
-        }
-
-        if (schedule.groups.length === newLocalSubjects.length) {
-          groupsToCompare = []
-        }
-
-        // Search groups until complete a schedule
-        while (groupsToCompare.length && schedule.groups.length < newLocalSubjects.length) {
-          if (groupsToCompare[0].nrc !== group.nrc
-            && !groupsToCompare[0].blocked
-            && !subjectsIdsUsed.has(groupsToCompare[0].subject.id)
-            && (allowFullGroups || groupsToCompare[0].quota.free > 0)) {
-            newMatrix = addToScheduleMatrix(groupsToCompare[0], schedule.matrix, matrixTemplate)
-            if (newMatrix) {
-              schedule.matrix = newMatrix
-              schedule.groups.push(groupsToCompare[0])
-              subjectsIdsUsed.add(groupsToCompare[0].subject.id)
-            }
-          }
-          groupsToCompare.shift()
-        }
-
-        if (schedule.groups.length < newLocalSubjects.length) {
-          break
-        }
-
-        schedule.groups.sort((a, b) => Number(b.nrc) - Number(a.nrc))
-        groupKey = schedule.groups.reduce((a, b) => a.nrc + b.nrc)
-
-        if (!groupKeys.has(groupKey)) {
-          schedules.push(schedule)
-          groupKeys.add(groupKey)
-        }
-      }
-    }
-
-    console.log('schedules', schedules.length)
-    if (schedules.length) setLocalCurrentSchedule(schedules[0])
-    setLocalSchedules(schedules)
-    setIsLoading(false)
-  }
-
-  const handleReset = onClean => {
+  const handleReset = () => {
     $('.ui-selected').map((_, elem) => elem.classList.remove('ui-selected'))
-    onClean(null)
+    setLocalMatrixTemplate(EMPTY_MATRIX)
+    loadSchedules(localSubjects, null)
+  }
+
+  const handleAllowGroups = (value) => {
+    setAllowFullGroups(value)
+    loadSchedules()
+  }
+
+  const handleOnStopSelecting = (matrix) => {
+    setLocalMatrixTemplate(matrix)
+    loadSchedules(localSubjects, matrix)
   }
 
   return (
@@ -216,17 +117,17 @@ function Home() {
         addSubject: newSubject => {
           const newLocalSubjects = [...localSubjects, newSubject]
           setLocalSubjects(newLocalSubjects)
-          generateSchedules(newLocalSubjects)
+          loadSchedules(newLocalSubjects)
         },
         removeSubject: index => {
           localSubjects.splice(index, 1)
-          generateSchedules()
+          loadSchedules()
         },
         updateGroupsStatus: (groups, blocked) => {
           groups.forEach(group => {
             group.blocked = blocked
           })
-          generateSchedules()
+          loadSchedules()
         },
       }}
     >
@@ -234,12 +135,7 @@ function Home() {
         value={{
           schedules: localSchedules,
           currentSchedule: localCurrentSchedule,
-          matrixTemplate: localMatrixTemplate,
           setCurrentSchedule: index => setLocalCurrentSchedule(localSchedules[index]),
-          setMatrixTemplate: matrix => {
-            setLocalMatrixTemplate(matrix)
-            generateSchedules(localSubjects, matrix)
-          },
         }}
       >
         <SubjectsContext.Consumer>
@@ -297,7 +193,7 @@ function Home() {
                   </SearchSection>
                   <MenuSection>
                     <SchedulesContext.Consumer>
-                      {({ schedules, setCurrentSchedule, setMatrixTemplate }) => (
+                      {({ schedules, setCurrentSchedule }) => (
                         <React.Fragment key="schedule">
                           <Row>
                             <Col xs={12} sm={12} md={12} lg={12}>
@@ -323,15 +219,12 @@ function Home() {
                                       <AllowFullGroups>
                                         <BlockCheckbox
                                           checked={allowFullGroups}
-                                          onChange={({ target }) => {
-                                            setAllowFullGroups(target.checked)
-                                            generateSchedules()
-                                          }}
+                                          onChange={({ target }) => handleAllowGroups(target.checked)}
                                         />
                                         <span>Permitir cursos sin cupo</span>
                                       </AllowFullGroups>
                                       <LinkuButton
-                                        onClick={() => handleReset(setMatrixTemplate)}
+                                        onClick={handleReset}
                                         color="#DA8686"
                                       >
                                         Limpiar filtro por horas
@@ -348,7 +241,7 @@ function Home() {
                           </Row>
                           <Row>
                             <Col xs={12} sm={12} md={12} lg={12}>
-                              <Table onStopSelecting={matrix => setMatrixTemplate(matrix)} />
+                              <Table onStopSelecting={handleOnStopSelecting} />
                               <ServerStatus />
                             </Col>
                           </Row>
