+++ src/pages/home/ServerStatus/ServerStatus.styles.js
@@ -1,0 +1,10 @@
+import styled, { createGlobalStyle } from 'styled-components'
+
+export const GlobalStyle = createGlobalStyle`
+.server_status {
+  font-size: 12px;
+  color: #082147;
+  margin-top: 14px;
+  float: right;
+}
+`
